feat(check): one static-literal resolver for every lane
static_literal_of resolves a whole-string bare ${{ <authority>.<name> }}
over the three static value authorities (const · inputs · config) to the
literal its declaration carries — an untyped value, or a typed
declaration's literal default. It is now THE resolver: the cost lane's
static_vars_array_len delegates to it (for_each fan-out counts), the
read-path lint's static_string_arg delegates to it (nika:read args), and
it exports publicly so the CLI's MODELS rung can judge a templated
model's declared default instead of skipping it wholesale — the trigger
named in models_rung.rs on 2026-07-29 (a third private copy is how two
lanes drift: measured that day, cost resolved ${{ const.model }}-class
refs while MODELS skipped them).

Behavioral tightening carried by the unification: static_string_arg's
old guard accepted any name without '.' or '[' — a hyphenated
${{ const.bad-name }} (outside the expression identifier grammar, and
the reject cost.rs already pins by mutation test) could resolve a read
path. The one resolver keeps the strict bare-ident guard everywhere;
the conservative direction (fewer claims, never wrong ones).

The public-api baseline gains exactly the one exported fn (regenerated
with the pinned 0.51.0; every pre-existing line byte-identical locally,
so the mac regeneration is trustworthy for this crate).
