+++ crates/nika-check/src/walk.rs
@@ -158,22 +158,31 @@ pub(crate) fn visit_json(value: &serde_json::Value, visit: &mut dyn FnMut(&str))
     }
 }
 
-/// Resolve an invoke arg to a STATIC string when it is a plain literal
-/// or a single `${{ const.X }}` / `${{ inputs.X }}` / `${{ config.X }}`
-/// whose declaration carries a literal value — the shapes a scaffold
-/// ships. Anything dynamic (task refs · concatenations) resolves to
-/// `None`: analysis never guesses.
-fn static_string_arg(wf: &RawWorkflow, value: &serde_json::Value) -> Option<String> {
-    let s = value.as_str()?;
-    let trimmed = s.trim();
-    if !trimmed.contains("${{") {
-        return Some(trimmed.to_owned());
-    }
-    let inner = trimmed.strip_prefix("${{")?.strip_suffix("}}")?.trim();
+/// THE shared static-value resolver: a whole-string bare
+/// `${{ <authority>.<name> }}` over the three value authorities whose
+/// declared value is static (`const.` · `inputs.` · `config.`), resolved
+/// to the LITERAL it declares — an untyped value, or a typed
+/// declaration's literal `default:`. One resolver, every lane: the cost
+/// ceiling counts `for_each` fan-outs through it, the read-path lint
+/// resolves `nika:read` args through it, and the CLI's MODELS rung
+/// judges a templated `model:`'s declared default through it. A private
+/// per-lane copy is how lanes drift (measured 2026-07-29: the cost lane
+/// resolved `${{ const.model }}`-class refs while MODELS skipped them
+/// wholesale — same expression, two verdicts).
+///
+/// `None` for everything else — navigation (`.field` · `[0]`),
+/// operators, concatenations, task refs, a name outside the expression
+/// identifier grammar (`[A-Za-z0-9_]`), or a typed declaration with no
+/// `default:` — which stays statically unknown: analysis never guesses.
+#[must_use]
+pub fn static_literal_of<'w>(wf: &'w RawWorkflow, expr: &str) -> Option<&'w serde_json::Value> {
+    let inner = expr.trim().strip_prefix("${{")?.strip_suffix("}}")?.trim();
     let (authority, name) = ["const.", "inputs.", "config."]
         .into_iter()
         .find_map(|ns| inner.strip_prefix(ns).map(|n| (ns, n)))?;
-    if name.contains(['.', '[']) {
+    // A BARE `<authority>.<ident>` only — further navigation or
+    // operators mean the runtime value is not the declared literal.
+    if name.is_empty() || !name.bytes().all(|b| b.is_ascii_alphanumeric() || b == b'_') {
         return None;
     }
     let block = match authority {
@@ -181,16 +190,27 @@ fn static_string_arg(wf: &RawWorkflow, value: &serde_json::Value) -> Option<Stri
         "inputs." => &wf.inputs,
         _ => &wf.config,
     };
-    block.iter().find_map(|(decl_name, decl)| {
-        if decl_name.value != name {
-            return None;
-        }
-        let default = match decl {
-            VarDecl::Untyped(v) => Some(v),
-            VarDecl::Typed { default, .. } => default.as_ref(),
-        };
-        default.and_then(|d| d.as_str()).map(str::to_owned)
-    })
+    let (_, decl) = block.iter().find(|(k, _)| k.value == name)?;
+    match decl {
+        VarDecl::Untyped(v)
+        | VarDecl::Typed {
+            default: Some(v), ..
+        } => Some(v),
+        VarDecl::Typed { default: None, .. } => None,
+    }
+}
+
+/// Resolve an invoke arg to a STATIC string when it is a plain literal
+/// or a bare authority ref [`static_literal_of`] resolves to a string
+/// literal — the shapes a scaffold ships. Anything dynamic (task refs ·
+/// concatenations) resolves to `None`: analysis never guesses.
+fn static_string_arg(wf: &RawWorkflow, value: &serde_json::Value) -> Option<String> {
+    let s = value.as_str()?;
+    let trimmed = s.trim();
+    if !trimmed.contains("${{") {
+        return Some(trimmed.to_owned());
+    }
+    static_literal_of(wf, trimmed)?.as_str().map(str::to_owned)
 }
 
 /// Every `nika:read` whose `path` arg resolves STATICALLY — the pure
